v0.130.0: harden migrations and writeback APIs
