+++ docker-compose.yml
@@ -26,7 +26,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    image: ghcr.io/trickle-labs/pg_ripple:0.129.0
+    image: ghcr.io/trickle-labs/pg_ripple:0.130.0
     environment:
       # S13-07 (v0.86.0): use POSTGRES_PASSWORD_FILE instead of plain-text password.
       # Generate: mkdir -p secrets && openssl rand -base64 32 > secrets/pg_password.txt
@@ -49,7 +49,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    image: ghcr.io/trickle-labs/pg_ripple:0.129.0
+    image: ghcr.io/trickle-labs/pg_ripple:0.130.0
     entrypoint: ["/usr/local/bin/pg_ripple_http"]
     ports:
       - "7878:7878"
@@ -73,4 +73,3 @@ secrets:
 
 volumes:
   pgdata:
-
