+++ roadmap/v0.130.0.md
@@ -1,0 +1,41 @@
+# v0.130.0 — Installation and migration integrity
+
+## Objective
+
+Make fresh installation and upgrades independently verifiable and impossible to
+bypass with a vacuous migration test.
+
+## Deliverables
+
+- [x] Parse the migration graph independently from `pg_ripple.control`, reject
+  malformed, duplicate, cyclic, ambiguous, gapped, or unreachable paths, and
+  emit `target/migration-graph.json`.
+- [x] Run the complete dynamic migration chain with representative data,
+  permissions, current-version, and latest writeback assertions.
+- [x] Emit normalized schema fingerprints and compare fresh-install and upgrade
+  results with an explicit allowlist for intentional differences.
+- [x] Add the `test-migration-full` command with a 30-minute timeout.
+- [x] Split oversized maintenance, Datalog parser/compiler, and JSON mapping
+  modules without changing their public SQL behavior.
+- [x] Add the validated `configure_json_writeback()` and `writeback_inspect()`
+  SQL APIs plus the HTTP configuration endpoint.
+- [x] Add module-size, package-installation, and upgrade-recovery checks.
+- [x] Publish the pre-GA upgrade support policy in the compatibility guide.
+- [x] Add the required `0.129.0 → 0.130.0` migration script.
+
+## Support policy
+
+Before v1.0.0, CI runs the complete historical sequential chain and release
+qualification covers the six most recent pre-GA release artifacts. After v1.0.0, direct upgrades are supported from the two previous
+minor releases; older supported versions upgrade sequentially through the
+migration graph.
+
+## Verification
+
+```text
+cargo fmt --all -- --check
+cargo clippy --features pg18 -- -D warnings
+cargo pgrx test pg18
+cargo pgrx regress pg18 --postgresql-conf "allow_system_table_mods=on"
+just test-migration-full
+```
