+++ src/datalog/compiler/aggregate.rs
@@ -1,0 +1,3 @@
+//! Aggregate emitter boundary.
+
+pub use super::sql::compile_aggregate_rule;
