+++ src/datalog/compiler/emit_cte.rs
@@ -1,0 +1,3 @@
+//! On-demand CTE emitter boundary.
+
+pub use super::sql::compile_on_demand_cte;
